hello-world/server/requirements.txt
mcp-python==0.1.4
requests>=2.31.0
